Fixed small bug concerning float matrices with WARMUP_RNG.
git-svn-id: file://localhost/home/tolik/svnrepo/branches/perf_tests@6622 c5418bc3-7546-4aae-8bad-fa6f2f3e53e6

+++ modules/ts/src/ts_perf.cpp
@@ -12,7 +12,7 @@ void randu(cv::Mat& m)
     }
     else if (m.depth() == CV_32F)
     {
-        double minmax[] = {-FLT_MAX, FLT_MAX};
+        float minmax[] = {-FLT_MAX, FLT_MAX};
         cv::Mat mr = m.reshape(1);
         cv::randu(mr, cv::Mat(1, 1, CV_32F, minmax), cv::Mat(1, 1, CV_32F, minmax + 1));
     }
